+++ feedback_grape/utils/states.py
@@ -14,7 +14,7 @@ def basis(n, k=0):
     """
     Basis state in n-dimensional Hilbert space.
     """
-    one_hot = jax.nn.one_hot(k, n, dtype=jnp.complex64)
+    one_hot = jax.nn.one_hot(k, n, dtype=jnp.complex128)
     return one_hot.reshape(n, 1)
